api/requirements.txt
fastapi
uvicorn
pydantic
httpx
python-Levenshtein
fuzzywuzzy
transformers
torch
finnhub-python
hf_xet
pytz